+++ source/renderer/assets/tpl/dialog/preferences.handlebars.js
@@ -43,6 +43,8 @@ module.exports = {"1":function(container,depth0,helpers,partials,data,blockParam
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.export.title",{"name":"i18n","hash":{},"data":data}))
     + "</a></li>\n        <li><a href=\"#prefs-tabs-zkn\">"
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.zkn.title",{"name":"i18n","hash":{},"data":data}))
+    + "</a></li>\n        <li><a href=\"#prefs-tabs-display\">"
+    + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.display.title",{"name":"i18n","hash":{},"data":data}))
     + "</a></li>\n        <li><a href=\"#prefs-tabs-advanced\">"
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.advanced",{"name":"i18n","hash":{},"data":data}))
     + "</a></li>\n      </ul>\n      <!-- General settings -->\n      <div id=\"prefs-tabs-general\">\n        <label for=\"app-lang\">"
@@ -171,7 +173,17 @@ module.exports = {"1":function(container,depth0,helpers,partials,data,blockParam
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.zkn.var_minute",{"name":"i18n","hash":{},"data":data}))
     + "</li>\n              <li><strong>%s</strong>: "
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.zkn.var_second",{"name":"i18n","hash":{},"data":data}))
-    + "</li>\n            </ul>\n          </p>\n        </div>\n      </div>\n      <!-- Export/seldomly used options -->\n      <div id=\"prefs-tabs-advanced\">\n        <p>"
+    + "</li>\n            </ul>\n          </p>\n        </div>\n      </div>\n      <!-- Display options -->\n      <div id=\"prefs-tabs-display\">\n        <p>"
+    + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.display.info",{"name":"i18n","hash":{},"data":data}))
+    + "</p>\n        <div class=\"box-left\">\n          <label for=\"imageWidth\">"
+    + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.display.image_width",{"name":"i18n","hash":{},"data":data}))
+    + "</label>\n          <input type=\"range\" min=\"1\" max=\"100\" id=\"imageWidth\" name=\"display.imageWidth\" value=\""
+    + alias3(alias4(((stack1 = (depth0 != null ? depth0.display : depth0)) != null ? stack1.imageWidth : stack1), depth0))
+    + "\">\n          <label for=\"imageHeight\">"
+    + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.display.image_height",{"name":"i18n","hash":{},"data":data}))
+    + "</label>\n          <input type=\"range\" min=\"1\" max=\"100\" id=\"imageHeight\" name=\"display.imageHeight\" value=\""
+    + alias3(alias4(((stack1 = (depth0 != null ? depth0.display : depth0)) != null ? stack1.imageHeight : stack1), depth0))
+    + "\">\n        </div>\n        <div class=\"box-right\">\n        </div>\n      </div>\n      <!-- Export/seldomly used options -->\n      <div id=\"prefs-tabs-advanced\">\n        <p>"
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.attachments_info",{"name":"i18n","hash":{},"data":data}))
     + "</p>\n        <textarea id=\"attachmentExtensions\" name=\"attachmentExtensions\">"
     + alias3(((helper = (helper = helpers.attachmentExtensions || (depth0 != null ? depth0.attachmentExtensions : depth0)) != null ? helper : alias2),(typeof helper === alias5 ? helper.call(alias1,{"name":"attachmentExtensions","hash":{},"data":data}) : helper)))